refactor(add to cart): move add to cart into cart actions directory
